# docs/swagger.yaml
basePath: /
definitions:
...
  keerja-backend_internal_dto_request.AdminDeleteCompanyRequest:
    properties:
      force:
        description: Force delete even with active jobs
        type: boolean
      reason:
        description: Reason for deletion
        maxLength: 500
        minLength: 10
        type: string
    required:
    - reason
    type: object
...
  keerja-backend_internal_dto_request.AdminUpdateCompanyRequest:
    properties:
      about:
        maxLength: 5000
        type: string
      benefits:
        items:
          type: string
        type: array
      company_name:
        maxLength: 200
        minLength: 2
        type: string
      company_size_id:
        type: integer
      culture:
        maxLength: 5000
        type: string
      description:
        maxLength: 2000
        type: string
      district_id:
        type: integer
      email_domain:
        maxLength: 100
        type: string
      full_address:
        maxLength: 500
        type: string
      industry_id:
        type: integer
      is_active:
        type: boolean
      latitude:
        maximum: 90
        minimum: -90
        type: number
      legal_name:
        maxLength: 200
        type: string
      longitude:
        maximum: 180
        minimum: -180
        type: number
      phone:
        maxLength: 30
        minLength: 10
        type: string
      postal_code:
        maxLength: 10
        type: string
      registration_number:
        maxLength: 100
        type: string
      verified:
        type: boolean
      website_url:
        type: string
    type: object
  keerja-backend_internal_dto_request.AdminUpdateCompanyStatusRequest:
    properties:
      grant_badge:
        description: Grant verification badge
        type: boolean
      notes:
        maxLength: 2000
        type: string
      rejection_reason:
        maxLength: 1000
        type: string
      status:
        enum:
        - pending_verification
        - verified
        - rejected
        - suspended
        - blacklisted
        type: string
    required:
    - status
    type: object
...
  keerja-backend_internal_dto_response.AdminCompaniesListResponse:
    properties:
      companies:
        items:
          $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyListItemResponse'
        type: array
      meta:
        $ref: '#/definitions/keerja-backend_internal_dto_response.PaginationMeta'
    type: object
  keerja-backend_internal_dto_response.AdminCompanyCreatorResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      role:
        description: owner, admin, recruiter
        type: string
      user_id:
        type: integer
    type: object
  keerja-backend_internal_dto_response.AdminCompanyDeleteResponse:
    properties:
      company_name:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: integer
      id:
        type: integer
      message:
        type: string
      reason:
        type: string
    type: object
  keerja-backend_internal_dto_response.AdminCompanyDetailResponse:
    properties:
      about:
        type: string
      banner_url:
        type: string
      benefits:
        items:
          type: string
        type: array
      company_name:
        type: string
      company_size:
        $ref: '#/definitions/keerja-backend_internal_dto_response.MasterCompanySizeResponse'
      company_size_id:
        type: integer
      company_type:
        type: string
      country:
        type: string
      created_at:
        description: Timestamps
        type: string
      creator_info:
        allOf:
        - $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyCreatorResponse'
        description: Owner/Creator Info
      culture:
        type: string
      description:
        description: Content
        type: string
      documents:
        description: Documents
        items:
          $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyDocumentResponse'
        type: array
      email_domain:
        type: string
      employer_users:
        description: Employer Users (Team members)
        items:
          $ref: '#/definitions/keerja-backend_internal_dto_response.AdminEmployerUserResponse'
        type: array
      full_address:
        type: string
      id:
        description: Basic Info
        type: integer
      industry:
        description: Legacy fields
        type: string
      industry_id:
        description: Master Data Relations
        type: integer
      industry_info:
        $ref: '#/definitions/keerja-backend_internal_dto_response.MasterIndustryResponse'
      is_active:
        type: boolean
      latitude:
        type: number
      legal_name:
        type: string
      location_info:
        $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyLocationResponse'
      logo_url:
        description: Media
        type: string
      longitude:
        type: number
      phone:
        type: string
      postal_code:
        type: string
      profile:
        allOf:
        - $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyProfileResponse'
        description: Profile
      registration_number:
        type: string
      size_category:
        type: string
      slug:
        type: string
      stats:
        allOf:
        - $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyStatsResponse'
        description: Statistics
      updated_at:
        type: string
      uuid:
        type: string
      verification_detail:
        allOf:
        - $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyVerificationDetailResponse'
        description: Verification Detail
      verified:
        description: Verification & Status
        type: boolean
      verified_at:
        type: string
      verified_by:
        type: integer
      verifier_info:
        $ref: '#/definitions/keerja-backend_internal_dto_response.AdminUserInfoResponse'
      website_url:
        description: Contact & Address
        type: string
    type: object
  keerja-backend_internal_dto_response.AdminCompanyDocumentResponse:
    properties:
      created_at:
        type: string
      document_name:
        type: string
      document_number:
        type: string
      document_type:
        type: string
      expiry_date:
        type: string
      file_path:
        type: string
      file_url:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_expired:
        type: boolean
      issue_date:
        type: string
      rejection_reason:
        type: string
      status:
        description: pending, approved, rejected, expired
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: integer
      uploaded_by_name:
        type: string
      verified_at:
        type: string
      verified_by:
        type: integer
      verified_by_name:
        type: string
    type: object
  keerja-backend_internal_dto_response.AdminCompanyListItemResponse:
    properties:
      active_jobs:
        type: integer
      company_name:
        type: string
      company_size:
        type: string
      created_at:
        type: string
      email:
        description: From owner/creator
        type: string
      id:
        type: integer
      industry:
        type: string
      is_active:
        type: boolean
      legal_name:
        type: string
      location:
        description: City, Province
        type: string
      phone:
        type: string
      registration_number:
        type: string
      reviewed_at:
        type: string
      slug:
        type: string
      total_applications:
        type: integer
      total_followers:
        type: integer
      total_jobs:
        description: Stats (quick view)
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
      verification_status:
        description: Verification info
        type: string
      verified:
        type: boolean
      verified_at:
        type: string
      verified_by:
        type: integer
      verifier_name:
        type: string
    type: object
...
  keerja-backend_internal_dto_response.AdminCompanyStatusUpdateResponse:
    properties:
      badge_granted:
        type: boolean
      company_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      rejection_reason:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: integer
      reviewed_by_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      verification_status:
        type: string
      verified:
        type: boolean
      verified_at:
        type: string
    type: object
...
  keerja-backend_internal_dto_response.AdminEmployerUserResponse:
    properties:
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_login:
        type: string
      position_title:
        type: string
      role:
        description: owner, admin, recruiter, viewer
        type: string
      user_id:
        type: integer
    type: object
...
  keerja-backend_internal_dto_response.MasterCompanySizeResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      max_employees:
        type: integer
      min_employees:
        type: integer
    type: object
...
  keerja-backend_internal_dto_response.PaginationMeta:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  keerja-backend_internal_utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@keerja.com
    name: API Support
  description: Job platform backend API with authentication, job management, applications,
    companies, and push notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Keerja Backend API
  version: "1.0"
paths:
...
  /api/v1/admin/companies:
    get:
      consumes:
      - application/json
      description: Get paginated list of companies with search, filter, and sort
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by company name, email, legal name
        in: query
        name: search
        type: string
      - description: Filter by verification status
        in: query
        name: status
        type: string
      - description: Filter by verified status
        in: query
        name: verified
        type: boolean
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Filter by industry ID
        in: query
        name: industry_id
        type: integer
      - description: Filter by company size ID
        in: query
        name: company_size_id
        type: integer
      - description: Filter by province ID
        in: query
        name: province_id
        type: integer
      - description: Filter by city ID
        in: query
        name: city_id
        type: integer
      - description: Filter by created date from (YYYY-MM-DD)
        in: query
        name: created_from
        type: string
      - description: Filter by created date to (YYYY-MM-DD)
        in: query
        name: created_to
        type: string
      - description: Sort by field
        enum:
        - company_name
        - created_at
        - verified_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompaniesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: List companies for moderation
      tags:
      - Admin - Company
  /api/v1/admin/companies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete company with validation. Cannot delete if has active jobs
        unless force flag is used.
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Force delete even with active jobs
        in: query
        name: force
        type: boolean
      - description: Delete request with reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_dto_request.AdminDeleteCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete company
      tags:
      - Admin - Company
    get:
      consumes:
      - application/json
      description: Get full company details including verification, documents, and
        stats
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get company detail
      tags:
      - Admin - Company
    put:
      consumes:
      - application/json
      description: Admin can edit company information to fix errors or typos
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_dto_request.AdminUpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update company details
      tags:
      - Admin - Company
  /api/v1/admin/companies/{id}/status:
    patch:
      consumes:
      - application/json
      description: Approve, reject, suspend, or blacklist a company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_dto_request.AdminUpdateCompanyStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_dto_response.AdminCompanyStatusUpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update company status
      tags:
      - Admin - Company
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"